+++ context/app/static/js/components/detailPage/BulkDataTransfer/index.ts
@@ -1,0 +1,3 @@
+import BulkDataTransfer from './BulkDataTransferSection';
+
+export default BulkDataTransfer;