+++ context/app/static/js/components/detailPage/Protocol/style.ts
@@ -1,0 +1,8 @@
+import { styled } from '@mui/material/styles';
+import Paper from '@mui/material/Paper';
+
+const StyledPaper = styled(Paper)({
+  padding: '20px 40px 20px 40px',
+});
+
+export { StyledPaper };